requirements.txt
flask
flask-jwt-extended
flask-sqlalchemy
flask-marshmallow
psycopg2-binary